# rpc/api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/execute": {
      "post": {
        "operationId": "GetProof",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetProofResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiGetProofRequest"
            }
          }
        ],
        "tags": [
          "Post"
        ]
      }
    },
    "/v1/info": {
      "get": {
        "operationId": "GetInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetInfoResponse"
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/v1/initialize": {
      "post": {
        "operationId": "Initialize",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiInitializeResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiInitializeRequest"
            }
          }
        ],
        "tags": [
          "Post"
        ]
      }
    },
    "/v1/reset": {
      "post": {
        "operationId": "Reset",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiResetResponse"
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/v1/shutdown": {
      "post": {
        "operationId": "Shutdown",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiShutdownResponse"
            }
          }
        },
        "tags": [
          "Post"
        ]
      }
    }
  },
  "definitions": {
    "apiExecuteRequest": {
      "type": "object",
      "properties": {
        "challenge": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "apiExecuteResponse": {
      "type": "object",
      "properties": {
        "proof": {
          "$ref": "#/definitions/apiProof"
        }
      }
    },
    "apiGetInfoResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        },
        "params": {
          "$ref": "#/definitions/apiParams"
        },
        "state": {
          "$ref": "#/definitions/apiState"
        }
      }
    },
    "apiGetProofRequest": {
      "type": "object",
      "properties": {
        "challenge": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "apiGetProofResponse": {
      "type": "object",
      "properties": {
        "proof": {
          "$ref": "#/definitions/apiProof"
        }
      }
    },
    "apiInitializeRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "apiInitializeResponse": {
      "type": "object",
      "properties": {
        "proof": {
          "$ref": "#/definitions/apiProof"
        }
      }
    },
    "apiParams": {
      "type": "object",
      "properties": {
        "spacePerUnit": {
          "type": "string",
          "format": "int64"
        },
        "difficulty": {
          "type": "integer",
          "format": "int32"
        },
        "t": {
          "type": "integer",
          "format": "int32"
        },
        "cacheLayer": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "apiProof": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "byte"
        },
        "challenge": {
          "type": "string",
          "format": "byte"
        },
        "merkleRoot": {
          "type": "string",
          "format": "byte"
        },
        "provenLeaves": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "byte"
          }
        },
        "proofNodes": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "byte"
          }
        }
      }
    },
    "apiResetResponse": {
      "type": "object",
      "properties": {
        "deletedDir": {
          "type": "string"
        },
        "numOfDeletedFiles": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "apiShutdownResponse": {
      "type": "object"
    },
    "apiState": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "byte"
        },
        "dir": {
          "type": "string"
        },
        "provenChallenges": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}